Break Mars generic text editor, in order to load huge source codes
The python image to assembly code converter is generating quite detailed bitmaps/codes. This results in a great # of LOC.

The assembler/compiler seems to deal fine with these aumount of LOC, but two embeded text editors completely freezes when trying to even load a single image's code.

So, in this paths, the Mars source was downloaded, and the code of the GENERIC TEXT EDITOR was change (just some lines were brutally deleted).

The text editor doens't work anymore, but the assembler and the bitmap are working just fine.